Rename toolkit-specific 'context' modules to 'event_loop'
