+++ traits_futures/asyncio/tests/test_context.py
@@ -15,7 +15,7 @@ Tests for the asyncio GUI context.
 
 import unittest
 
-from traits_futures.asyncio.context import AsyncioEventLoop
+from traits_futures.asyncio.event_loop import AsyncioEventLoop
 from traits_futures.tests.i_event_loop_tests import IEventLoopTests
