+++ traits_futures/asyncio/tests/test_pingee.py
@@ -14,7 +14,7 @@ Tests for the asyncio implementations of IPingee and IPinger.
 
 import unittest
 
-from traits_futures.asyncio.context import AsyncioEventLoop
+from traits_futures.asyncio.event_loop import AsyncioEventLoop
 from traits_futures.testing.gui_test_assistant import GuiTestAssistant
 from traits_futures.tests.i_pingee_tests import IPingeeTests
