+++ traits_futures/qt/tests/test_context.py
@@ -23,6 +23,6 @@ from traits_futures.tests.i_event_loop_tests import IEventLoopTests
 class TestQtEventLoop(IEventLoopTests, unittest.TestCase):
     def context_factory(self):
         """Factory for instances of the context."""
-        from traits_futures.qt.context import QtEventLoop
+        from traits_futures.qt.event_loop import QtEventLoop
 
         return QtEventLoop()
