+++ traits_futures/qt/tests/test_event_loop_helper.py
@@ -24,6 +24,6 @@ from traits_futures.tests.i_event_loop_helper_tests import (
 class TestEventLoopHelper(IEventLoopHelperTests, unittest.TestCase):
     def event_loop_factory(self):
         """Create a suitable IEventLoop instance."""
-        from traits_futures.qt.context import QtEventLoop
+        from traits_futures.qt.event_loop import QtEventLoop
 
         return QtEventLoop()
