+++ traits_futures/wx/tests/test_pingee.py
@@ -22,6 +22,6 @@ from traits_futures.tests.i_pingee_tests import IPingeeTests
 @requires_wx
 class TestPingee(GuiTestAssistant, IPingeeTests, unittest.TestCase):
     def event_loop_factory(self):
-        from traits_futures.wx.context import WxEventLoop
+        from traits_futures.wx.event_loop import WxEventLoop
 
         return WxEventLoop()
